added old Questionnaire migrations.
the issue is that migrations dynamically specify table names,
EG, Question.table_name. Since we are adding a namespace,
Qe::Question, the Question constant is not recognized.

Therefore, my vote is include these old migration files as
reference files, but create new namespaced migrations and
note this in a changelog.